requirements.txt
lxml
statscounter
Pillow
libextract
requests
rdflib
python-docx
bs4
